Make tests not TLS specific
